requirements.txt
telethon==1.34.0
apscheduler==3.10.4
python-dotenv==1.0.0
